+++ aesara/link/numba/dispatch/elemwise.py
@@ -38,8 +38,13 @@ from aesara.scalar.basic import (
 from aesara.scalar.basic import add as add_as
 from aesara.scalar.basic import scalar_maximum
 from aesara.tensor.elemwise import CAReduce, DimShuffle, Elemwise
-from aesara.tensor.math import MaxAndArgmax, MulWithoutZeros
-from aesara.tensor.nnet.basic import LogSoftmax, Softmax, SoftmaxGrad
+from aesara.tensor.math import (
+    LogSoftmax,
+    MaxAndArgmax,
+    MulWithoutZeros,
+    Softmax,
+    SoftmaxGrad,
+)
 
 
 @singledispatch
